chore: rename template files
