+++ packages/cli/templates/react/igr-ts/bullet-graph/default/files/src/app/__path__/__filePrefix__.tsx
