+++ packages/cli/templates/react/igr-ts/bullet-graph/default/files/src/app/__path__/style.module.css
