+++ packages/cli/templates/react/igr-ts/category-chart/default/files/src/app/__path__/style.module.css
