+++ packages/cli/templates/react/igr-ts/doughnut-chart/default/files/src/app/__path__/style.module.css
