+++ packages/cli/templates/react/igr-ts/financial-chart/default/files/src/app/__path__/style.module.css
