+++ packages/cli/templates/react/igr-ts/grid/basic/files/src/app/__path__/data.tsx
