+++ packages/cli/templates/react/igr-ts/linear-gauge/default/files/src/app/__path__/__filePrefix__.tsx
