+++ packages/cli/templates/react/igr-ts/linear-gauge/default/files/src/app/__path__/style.module.css
