+++ packages/cli/templates/react/igr-ts/pie-chart/default/files/src/app/__path__/style.module.css
