--- packages/cli/templates/react/igr-ts/projects/_base/files/src/App.tsx
@@ -1,10 +1,0 @@
-import { BrowserRouter } from "react-router-dom";
-
-export default function App() {
-
-	return (
-    <div className="app">
-      <BrowserRouter></BrowserRouter>
-    </div>
-	)
-}
