+++ packages/cli/templates/react/igr-ts/projects/_base/files/src/app/App.test.tsx
