+++ packages/cli/templates/react/igr-ts/projects/_base/files/src/app/App.tsx
@@ -1,0 +1,6 @@
+export default function App() {
+
+	return (
+		<div className="app"></div>
+	)
+}
