+++ packages/cli/templates/react/igr-ts/projects/_base_with_home/files/src/app/App.test.tsx
