+++ packages/cli/templates/react/igr-ts/projects/top-nav/files/src/app/App.css
