+++ packages/cli/templates/react/igr-ts/projects/top-nav/files/src/app/app-routes.tsx
@@ -1,4 +1,4 @@
-import Home from './views/home';
+import Home from './home/home';
 
 export const routes = [
   { path: '/', element: <Home />, text: 'Home' }
