+++ packages/cli/templates/react/igr-ts/radial-gauge/default/files/src/app/__path__/__filePrefix__.tsx
